pyproject.toml
[project]
name = "simbolo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.54",
    "langchain-groq>=0.3.2",
    "langserve>=0.3.1",
    "pydantic>=2.11.3",
    "uvicorn>=0.34.2",
]
